pyproject.toml
[project]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = [
    {name = "Ailton teles xavier",email = "atxshared@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "fastapi[standard] (>=0.135.3,<0.136.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.49,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)"
]

[tool.poetry]
packages = [{include = "fast_zero", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.15.10,<0.16.0)",
    "typos (>=1.45.0,<2.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "freezegun (>=1.5.5,<2.0.0)",
    "testcontainers (>=4.14.2,<5.0.0)"
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.typos.files]
extend-exclude = ["*.md"]

[tool.taskipy.tasks]
pre_lint = 'typos'
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev src/core/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src -vv'
post_test = 'coverage html'

[tool.coverage.run]
core = "ctrace"
concurrency = ["thread", "greenlet"]

[tool.typos.default]
extend-ignore-identifiers-re = ["selectin"]